requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
gunicorn==19.9.0
pandas==0.24.2
psycopg2-binary=2.8.2